# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API Falabella FIF
  description: Esta API esta diseÃ±ada para ser una prueba para los nuevos candidatos
    al equipo.
  contact:
    email: lugaetea@falabella.cl
  version: 1.0.0
servers:
- url: /
tags:
- name: cerveza
  description: rica cerveza..
paths:
  /beers:
    get:
      tags:
      - cerveza
      summary: Lista todas las cervezas
      description: |
        Lista todas las cervezas que se encuentran en la base de datos
      operationId: searchBeers
      responses:
        200:
          description: Operacion exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BeerItem'
    post:
      tags:
      - cerveza
      summary: Ingresa una nueva cerveza
      description: Ingresa una nueva cerveza
      operationId: addBeers
      requestBody:
        description: Ingresa una nueva cerveza
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeerItem'
      responses:
        201:
          description: Cerveza creada
        400:
          description: Request invalida
        409:
          description: El ID de la cerveza ya existe
  /beers/{beerID}:
    get:
      tags:
      - cerveza
      summary: Lista todas las cervezas
      description: |
        Busca una cerveza por su Id
      operationId: searchBeerById
      parameters:
      - name: beerID
        in: path
        description: Busca una cerveza por su Id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        200:
          description: Operacion exitosa
        404:
          description: El Id de la cerveza no existe
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BeerItem'
components:
  schemas:
    BeerItem:
      required:
      - Brewery
      - Country
      - Id
      - Name
      - Price
      type: object
      properties:
        Id:
          type: integer
          example: 1
        Name:
          type: string
          example: Golden
        Brewery:
          type: string
          example: Kross
        Country:
          type: string
          example: Chile
        Price:
          type: number
          example: 10.5